FIXES
- статус пользователя устанавливается через состояние loggedIn
- редьюсеры и экшены разделены на два файла - для обычных действий и связанных с авторизацией
