+++ src/services/reducers/index.js
@@ -2,19 +2,21 @@ import { combineReducers } from 'redux';
 import {
   getAllIngredients,
   makeBurger,
-  chooseIngredient,
-  setOrder,
-  setPreloader,
-  setUserData,
-  setRestorePassword
+  setPreloader
 } from './reducers';
+import {
+  setOrder,
+  setUserData,
+  setLoggedIn,
+  setRestorePassword,
+} from './reducersWithAuth';
 
 export const rootReducer = combineReducers({
   ingredients: getAllIngredients,
   burger: makeBurger,
-  ingredient: chooseIngredient,
   order: setOrder,
   preloader: setPreloader,
   user: setUserData,
+  loggedIn: setLoggedIn,
   restore: setRestorePassword
 });
