+++ src/services/reducers/reducers.js
@@ -1,11 +1,7 @@
 import {
   initialIngredients,
   initialBurger,
-  initialIngredient,
-  initialOrder,
   initialPreloader,
-  initialUser,
-  initialRestore
 } from '../../utils/initialStates';
 import {
   GET_INGREDIENTS_REQUEST,
@@ -13,20 +9,8 @@ import {
   GET_INGREDIENTS_FAILED,
   SET_CURRENT_TAB,
   CHANGE_BURGER,
-  OPEN_SELECTED_INGREDIENT,
-  CLOSE_SELECTED_INGREDIENT,
-  MAKE_ORDER_REQUEST,
-  MAKE_ORDER_SUCCESS,
-  MAKE_ORDER_FAILED,
-  CLOSE_ORDER,
   ENABLE_PRELOADER,
   DISABLE_PRELOADER,
-  SET_USER_REQUEST,
-  SET_USER_SUCCESS,
-  SET_USER_TOKEN_SUCCESS,
-  SET_USER_FAILED,
-  RESTORE_STARTED,
-  RESTORE_FINISHED
 } from '../actions/index';
 
 function getAllIngredients(state = initialIngredients, action) {
@@ -71,60 +55,6 @@ function makeBurger(state = initialBurger, action) {
   }
 };
 
-function chooseIngredient(state = initialIngredient, action) {
-  switch (action.type) {
-    case OPEN_SELECTED_INGREDIENT: {
-      return {
-        selectedIngredient: action.selectedIngredient,
-        open: true
-      };
-    }
-    case CLOSE_SELECTED_INGREDIENT: {
-      return {
-        selectedIngredient: {},
-        open: false
-      };
-    }
-    default: { return state; }
-  }
-};
-
-function setOrder(state = initialOrder, action) {
-  switch (action.type) {
-    case MAKE_ORDER_REQUEST: {
-      return {
-        ...state,
-        orderRequest: true
-      };
-    }
-    case MAKE_ORDER_SUCCESS: {
-      return {
-        ...state,
-        order: action.order,
-        orderRequest: false,
-        orderFailed: false,
-        open: true
-      };
-    }
-    case MAKE_ORDER_FAILED: {
-      return {
-        ...state,
-        orderRequest: false,
-        orderFailed: true
-      };
-    }
-    case CLOSE_ORDER: {
-      return {
-        order: {},
-        orderRequest: false,
-        orderFailed: false,
-        open: false
-      };
-    }
-    default: { return state; }
-  }
-};
-
 function setPreloader(state = initialPreloader, action) {
   switch (action.type) {
     case ENABLE_PRELOADER: {
@@ -137,62 +67,8 @@ function setPreloader(state = initialPreloader, action) {
   }
 };
 
-function setUserData(state = initialUser, action) {
-  switch (action.type) {
-    case SET_USER_REQUEST: {
-      return {
-        ...state,
-        userRequest: true
-      };
-    }
-    case SET_USER_SUCCESS: {
-      return {
-        ...action.userData,
-        userRequest: false,
-        userFailed: false,
-      }
-    }
-    case SET_USER_TOKEN_SUCCESS: {
-      return {
-        ...state,
-        ...action.userData,
-        userRequest: false,
-        userFailed: false,
-      }
-    }
-    case SET_USER_FAILED: {
-      return {
-        ...state,
-        userRequest: false,
-        userFailed: true,
-      }
-    }
-    default: { return state; }
-  }
-};
-
-function setRestorePassword(state = initialRestore, action) {
-  switch (action.type) {
-    case RESTORE_STARTED: {
-      return {
-        restore: true
-      };
-    }
-    case RESTORE_FINISHED: {
-      return {
-        restore: false
-      }
-    }
-    default: { return state; }
-  }
-};
-
 export {
   getAllIngredients,
   makeBurger,
-  chooseIngredient,
-  setOrder,
-  setPreloader,
-  setUserData,
-  setRestorePassword
+  setPreloader
 };
